Refactor (backend): Auth 컨트롤러로 이동
